+++ libs/notitia-templates/src/templates/locales/nb-NO/index.ts
@@ -1,0 +1,13 @@
+import type { TemplateRegistry } from '../../../types';
+import { emailTemplatesNbNO } from './email';
+import { smsTemplatesNbNO } from './sms';
+
+/**
+ * Combined Norwegian Bokmål templates
+ */
+export const templatesNbNO: TemplateRegistry = {
+  ...emailTemplatesNbNO,
+  ...smsTemplatesNbNO,
+};
+
+export { emailTemplatesNbNO, smsTemplatesNbNO };